+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ##0.5.0 - 2014-02-11
  - Support protocol 5 - multiple devices.
 
+## 0.4.2 - 2014-03-12
+ - Allow controller.use to be called idempotently
+ - Fix issue where requestAnimationFrame would not be used in Chrome
+ - Exclude node code from browser js
+ - The plugin pipeline now runs on animationFrames rather than deviceFrames when in the browser.
+
 ##0.4.1 - 2014-02-11
  - Allow controller.use to accept a factory directly (node support)